use non-wrapping blocking composition for union projection.
If the function used in blocking composition does not introduce any new
'Resolvables', we do not need to to 'wrap-tree' (which is very expensive
for deeply nested structures).

In the union projection, the inputs to 'union-of-maps' are fully resolved,
so the output will be fully resolved as well. We do not need to rewrap
the tree.

This brings the mean of the deep union projection benchmark down to
~280ms from ~650ms, providing an improvement of over 50%.